Feat/tour execution tracking

+++ src/Modules/Tours/Explorer.Tours.Infrastructure/Database/Repositories/TourPointDbRepository.cs
@@ -23,9 +23,9 @@ namespace Explorer.Tours.Infrastructure.Database.Repositories
                          .ToList();
         }
 
-        public TourPoint Get(int id)
+        public TourPoint Get(long id)
         {
-            var p = _dbSet.Find((long)id);
+            var p = _dbSet.Find(id);
             if (p == null) throw new NotFoundException($"Point not found {id}");
             return p;
         }
